Add files via upload
Calculation of geostrophic currents based on satellite altimetry data.

Definition of geostrophic current:
The surface geostrophic velocity is proportional to the slope of the ocean surface topography. This allows indirectly to observe geostrophic circulation through the measurement of the ocean’s topography by satellites. 
The ocean can be considered as geostrophic balanced in an approximation. This means that the horizontal pressure gradient is equal to the Coriolis force under certain simplifications.
These simplifications include the following assumptions:

•	the flow has no acceleration
•	only the horizontal velocity is considered
•	the only external force is gravity
•	friction is negligible

Due to these simplifications, only two terms of the momentum equation remain relevant: the pressure gradient and the Coriolis force. The geostrophic approximation leads to a simple relation between surface slope and surface current.